"prediction api 구현완료"
