+++ app/schemas/predict.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+
+class ImagePredictionRequest(BaseModel):
+    image_url: str
+
+class PredictionResult(BaseModel):
+    tag_name: str
+    probability: float
